Refactor: remove /admin from API routes
